+++ src/components/spartan/SCombobox/index.ts
@@ -1,5 +1,0 @@
-export { default as SCombobox } from './SCombobox.vue';
-export { default as SComboboxOption } from './SComboboxOption.vue';
-export { default as SComboboxOptionGroup } from './SComboboxOptionGroup.vue';
-
-export type { TComboboxProps, TComboboxEmits } from './types';