+++ src/components/spartan/SSkeleton/types.ts
@@ -1,0 +1,9 @@
+import type { ClassNameValue } from 'tailwind-merge';
+
+export type TSkeletonProps = {
+    /**
+     * @en Additional CSS classes to apply to the skeleton element. Typically used to control width, height, and border radius.
+     * @es Clases CSS adicionales para aplicar al elemento skeleton. Tipicamente usadas para controlar ancho, alto y radio de borde.
+     */
+    class?: ClassNameValue;
+};